refactor: signin, signup and forgot password pages are made seperated
