+++ src/app/[locale]/(content)/forgot-password/page.tsx
@@ -1,6 +1,6 @@
-import SignupForm from "@/src/components/customs/Signup/SignupForm";
-import { Metadata } from "next";
+import ForgotPasswordForm from "@/src/components/customs/Signup/ForgotPasswordForm";
 import React from "react";
+import { Metadata } from "next";
 
 export async function generateMetadata({
   params: { locale },
@@ -12,21 +12,21 @@ export async function generateMetadata({
   switch (locale) {
     case "tr":
       return {
-        title: "Kayıt ol",
-        description: "Hesap oluşturmak için kayıt olun",
+        title: "Şifremi unuttum",
+        description: "Şifrenizi sıfırlayın",
       };
     default:
       return {
-        title: "Sign up",
-        description: "Sign up to create an account",
+        title: "Forgot password",
+        description: "Reset your password",
       };
   }
 }
 
-export default function Page() {
+export default function page() {
   return (
     <main className="absolute grid place-items-center w-full h-[calc(100%-64px)] p-2">
-      <SignupForm />
+      <ForgotPasswordForm />
     </main>
   );
 }
