+++ src/messages/tr.json
@@ -4,7 +4,16 @@
       "Word List": "Kelime Listesi",
       "Sign In": "Giriş Yap"
     },
-    "Signup":{
+    "SigninForm":{
+      "Sign in with Google": "Google ile giriş yap",
+      "Don't have an account?" : "Hesabınız yok mu?",
+      "Sign Up": "Kaydolun",
+      "Sign In": "Giriş Yap",
+      "Forgot Password": "Şifrenizi mi unuttunuz?",
+      "Password": "Şifre",
+      "Username or Email": "Kullanıcı Adı veya E-posta"
+    },
+    "SignupForm":{
       "Sign up with Google": "Google ile kaydol",
       "Sign in with Google": "Google ile giriş yap",
       "Create a new account": "Yeni hesap oluştur",
@@ -14,18 +23,20 @@
       "Password": "Şifre",
       "Confirm Password": "Şifreyi Onayla",
       "Sign Up Button": "Kaydol",
-      "Sign Up": "Kaydolun",
       "Already have an account?": "Zaten hesabınız var mı?",
       "Login": "Giriş Yapın",
-      "Username or Email": "Kullanıcı Adı veya E-posta",
-      "Don't have an account?" : "Hesabınız yok mu?",
-      "Sign In": "Giriş Yap",
       "Invalid username, email or password": "Geçersiz kullanıcı adı, e-posta veya şifre",
       "Account created successfully, please sign in" : "Hesap başarıyla oluşturuldu. Lütfen giriş yapın.",
-      "Forgot Password": "Şifrenizi mi unuttunuz?",
       "Send Email": "E-posta Gönder",
       "Email Sent": "Doğrulama e-postası gönderildi, lütfen e-postanızı kontrol edin.",
       "User created with google auth": "Google kimlik doğrulamasıyla kullanıcı oluşturulmuş. Lütfen Google ile giriş yapın."
+    },
+    "ForgotPasswordForm":{
+      "Forgot Password": "Şifremi Unuttum",
+      "Email": "E-posta",
+      "Send Email": "E-posta Gönder",
+      "Email Sent": "Doğrulama e-postası gönderildi, lütfen e-postanızı kontrol edin.",
+      "Login": "Giriş Yapın"
     },
     "ResetPassword": {
       "Password": "Şifre",
