+++ src/middleware.ts
@@ -3,7 +3,14 @@ import createIntlMiddleware from "next-intl/middleware";
 import { NextRequest } from "next/server";
 
 const locales = ["en", "tr"];
-const publicPages = ["/", "/search", "/signup", "/reset-password/(.*)"];
+const publicPages = [
+  "/",
+  "/search",
+  "/signup",
+  "/signin",
+  "/forgot-password",
+  "/reset-password/(.*)",
+];
 
 const intlMiddleware = createIntlMiddleware({
   locales,
